fix: plugin load

+++ cdps/plugin/manager.py
@@ -3,6 +3,7 @@ import json
 import os
 import shutil
 import sys
+import threading
 import zipfile
 
 from cdps.utils.logger import Log
@@ -44,7 +45,7 @@ class Manager:
 class Plugin():
     _instance = None
 
-    def __new__(cls, log: Log = None, event_manager=None):
+    def __new__(cls, log=None, event_manager=None):
         if not cls._instance:
             if log is None or event_manager is None:
                 raise ValueError(
@@ -56,10 +57,13 @@ class Plugin():
     def init(self, log: Log, event_manager):
         self.log = log
         self.event_manager = event_manager
-        self.loaded_plugins_list = []
-
         self.modules = {}
+        self.all_stopped = threading.Event()
+        self.lock = threading.Lock()
+        self.loaded_plugins_list = []
+        self._load_initial_plugins()
 
+    def _load_initial_plugins(self):
         for entry in os.listdir(directory_path):
             full_path = os.path.join(directory_path, entry)
             if os.path.isfile(full_path) and ".cdps" in full_path:
@@ -74,15 +78,11 @@ class Plugin():
         for entry in os.listdir(directory_path):
             full_path = os.path.join(directory_path, entry)
             if not "__" in full_path and not os.path.isfile(full_path):
-                if os.path.isfile(os.path.join(full_path, "main.py")):
-                    if os.path.isfile(os.path.join(full_path, "cdps.json")):
-                        all_plugins.append(entry)
-                    else:
-                        self.log.logger.error(
-                            "Plugin [ {} ] Load Failed (cdps.json)".format(entry))
+                if os.path.isfile(os.path.join(full_path, "main.py")) and os.path.isfile(os.path.join(full_path, "cdps.json")):
+                    all_plugins.append(entry)
                 else:
                     self.log.logger.error(
-                        "Plugin [ {} ] Load Failed (main.py)".format(entry))
+                        f"Plugin [ {entry} ] Load Failed (missing main.py or cdps.json)")
         return all_plugins
 
     def load_info(self, plugins_info, plugins_list):
@@ -115,19 +115,18 @@ class Plugin():
 
     def load_plugins(self, plugins_list):
         for plugin in plugins_list:
-            config_path = os.path.join("./config/", "{}.json".format(plugin))
+            config_path = os.path.join("./config/", f"{plugin}.json")
             full_path = os.path.join(directory_path, plugin)
-            if os.path.isfile(os.path.join(full_path, "config.json")):
-                if not os.path.isfile(config_path):
-                    self.log.logger.warning(
-                        "Plugin [ {} ] Config Generate".format(plugin))
-                    shutil.copy(os.path.join(
-                        full_path, "config.json"), config_path)
+            if os.path.isfile(os.path.join(full_path, "config.json")) and not os.path.isfile(config_path):
+                shutil.copy(os.path.join(
+                    full_path, "config.json"), config_path)
+                self.log.logger.warning(
+                    f"Plugin [ {plugin} ] Config Generated")
             self.__reload_module__(plugin, os.path.join(full_path, "main.py"))
-            self.log.logger.info("Plugin [ {} ] Loaded".format(plugin))
+            self.log.logger.info(f"Plugin [ {plugin} ] Loaded")
             self.loaded_plugins_list.append(plugin)
         return self.loaded_plugins_list
-    
+
     def reload_load_plugins(self, name):
         if name in self.loaded_plugins_list:
             config_path = os.path.join("./config/", "{}.json".format(name))
@@ -144,8 +143,30 @@ class Plugin():
             self.log.logger.error("Plugin [ {} ] Reload Failed".format(name))
 
     def __reload_module__(self, module_name, path_to_module):
-        spec = importlib.util.spec_from_file_location(
-            module_name, path_to_module)
-        module = importlib.util.module_from_spec(spec)
-        sys.modules[module_name] = module
-        spec.loader.exec_module(module)
+        if module_name in self.modules:
+            self.stop_module(module_name)
+
+        stop_event = threading.Event()
+
+        def load_and_run_module():
+            spec = importlib.util.spec_from_file_location(
+                module_name, path_to_module)
+            module = importlib.util.module_from_spec(spec)
+            sys.modules[module_name] = module
+            spec.loader.exec_module(module)
+            if hasattr(module, 'task'):
+                module.task(stop_event)
+        thread = threading.Thread(target=load_and_run_module)
+        thread.daemon = True
+        thread.start()
+        self.modules[module_name] = (thread, stop_event)
+
+    def stop_module(self, module_name):
+        if module_name in self.modules:
+            _, stop_event = self.modules[module_name]
+            stop_event.set()
+            self.log.logger.warning("Plugin [ {} ] Unloaded".format(module_name))
+
+    def stop_all_modules(self):
+        for _, (_, stop_event) in self.modules.items():
+            stop_event.set()
